rename mem_available to system_memory
