+++ src/libnetdata/os/os.h
@@ -7,7 +7,7 @@
 #include <sys/syscall.h>
 #endif
 
-#include "mem_available.h"
+#include "system_memory.h"
 #include "random.h"
 #include "timestamps.h"
 #include "setproctitle.h"
