+++ src/libnetdata/os/system_memory.c
